fixed WebDAV upload

+++ bar/chunks.c
@@ -2686,7 +2686,7 @@ Errors Chunk_close(ChunkInfo *chunkInfo)
         return error;
       }
       size = chunkInfo->io->getSize(chunkInfo->ioUserData);
-      if ((size >= 0LL) && (chunkInfo->offset+CHUNK_HEADER_SIZE+chunkInfo->size > size))
+      if ((size >= 0LL) && (chunkInfo->offset+CHUNK_HEADER_SIZE+chunkInfo->size > (uint64)size))
       {
         return ERROR_INVALID_CHUNK_SIZE;
       }
